+++ packages/plugin-inbox-ui/src/settings/channels/containers/Sidebar.tsx
@@ -1,24 +1,24 @@
-import * as compose from 'lodash.flowright';
+import * as compose from "lodash.flowright";
 
-import { Alert, confirm, withProps } from '@erxes/ui/src/utils';
+import { Alert, confirm, withProps } from "@erxes/ui/src/utils";
 import {
   IButtonMutateProps,
   IRouterProps,
-  MutationVariables
-} from '@erxes/ui/src/types';
-import { mutations, queries } from '../graphql';
+  MutationVariables,
+} from "@erxes/ui/src/types";
+import { mutations, queries } from "../graphql";
 
-import { AppConsumer } from 'coreui/appContext';
-import ButtonMutate from '@erxes/ui/src/components/ButtonMutate';
-import { ChannelsCountQueryResponse } from '../types';
-import { ChannelsQueryResponse } from '@erxes/ui-inbox/src/settings/channels/types';
-import React from 'react';
-import { RemovePipelineLabelMutationResponse } from '@erxes/ui-cards/src/boards/types';
-import Sidebar from '../components/Sidebar';
-import { gql } from '@apollo/client';
-import { graphql } from '@apollo/client/react/hoc';
-import inboxQueries from '@erxes/ui-inbox/src/inbox/graphql/queries';
-import { withRouter } from 'react-router-dom';
+import { AppConsumer } from "coreui/appContext";
+import ButtonMutate from "@erxes/ui/src/components/ButtonMutate";
+import { ChannelsCountQueryResponse } from "../types";
+import { ChannelsQueryResponse } from "@erxes/ui-inbox/src/settings/channels/types";
+import React from "react";
+import { RemovePipelineLabelMutationResponse } from "@erxes/ui-cards/src/boards/types";
+import Sidebar from "../components/Sidebar";
+import { gql } from "@apollo/client";
+import { graphql } from "@apollo/client/react/hoc";
+import inboxQueries from "@erxes/ui-inbox/src/inbox/graphql/queries";
+import { withRouter } from "react-router-dom";
 
 type Props = {
   queryParams: any;
@@ -41,24 +41,26 @@ const SidebarContainer = (props: FinalProps) => {
     queryParams,
     history,
     currentChannelId,
-    currentUserId
+    currentUserId,
   } = props;
 
   const channels = channelsQuery.channels || [];
   const channelsTotalCount = channelsCountQuery.channelsTotalCount || 0;
 
   // remove action
-  const remove = channelId => {
-    confirm().then(() => {
+  const remove = (channelId) => {
+    confirm("This will permanently delete are you absolutely sure?", {
+      hasDeleteConfirm: true,
+    }).then(() => {
       removeMutation({
-        variables: { _id: channelId }
+        variables: { _id: channelId },
       })
         .then(() => {
-          Alert.success('You successfully deleted a channel.');
+          Alert.success("You successfully deleted a channel.");
 
-          history.push('/settings/channels');
+          history.push("/settings/channels");
         })
-        .catch(error => {
+        .catch((error) => {
           Alert.error(error.message);
         });
     });
@@ -69,7 +71,7 @@ const SidebarContainer = (props: FinalProps) => {
     values,
     isSubmitted,
     callback,
-    object
+    object,
   }: IButtonMutateProps) => {
     return (
       <ButtonMutate
@@ -84,7 +86,7 @@ const SidebarContainer = (props: FinalProps) => {
         isSubmitted={isSubmitted}
         type="submit"
         successMessage={`You successfully ${
-          object ? 'updated' : 'added'
+          object ? "updated" : "added"
         } a ${name}`}
       />
     );
@@ -96,7 +98,7 @@ const SidebarContainer = (props: FinalProps) => {
     channelsTotalCount,
     remove,
     renderButton,
-    loading: channelsQuery.loading
+    loading: channelsQuery.loading,
   };
 
   return <Sidebar {...updatedProps} />;
@@ -111,26 +113,26 @@ const getRefetchQueries = (
     {
       query: gql(queries.channels),
       variables: {
-        perPage: queryParams.limit ? parseInt(queryParams.limit, 10) : 20
-      }
+        perPage: queryParams.limit ? parseInt(queryParams.limit, 10) : 20,
+      },
     },
     {
       query: gql(queries.channels),
-      variables: {}
+      variables: {},
     },
     {
       query: gql(queries.integrationsCount),
-      variables: {}
+      variables: {},
     },
     {
       query: gql(queries.channelDetail),
-      variables: { _id: currentChannelId || '' }
+      variables: { _id: currentChannelId || "" },
     },
     { query: gql(queries.channelsCount) },
     {
       query: gql(inboxQueries.channelList),
-      variables: { memberIds: [currentUserId] }
-    }
+      variables: { memberIds: [currentUserId] },
+    },
   ];
 };
 
@@ -139,29 +141,29 @@ const WithProps = withProps<Props>(
     graphql<Props, ChannelsQueryResponse, { perPage: number }>(
       gql(queries.channels),
       {
-        name: 'channelsQuery',
+        name: "channelsQuery",
         options: ({ queryParams }: { queryParams: any }) => ({
           variables: {
-            perPage: queryParams.limit ? parseInt(queryParams.limit, 10) : 20
+            perPage: queryParams.limit ? parseInt(queryParams.limit, 10) : 20,
           },
-          fetchPolicy: 'network-only'
-        })
+          fetchPolicy: "network-only",
+        }),
       }
     ),
     graphql<Props, ChannelsCountQueryResponse, {}>(gql(queries.channelsCount), {
-      name: 'channelsCountQuery'
+      name: "channelsCountQuery",
     }),
     graphql<Props, RemovePipelineLabelMutationResponse, MutationVariables>(
       gql(mutations.channelRemove),
       {
-        name: 'removeMutation',
+        name: "removeMutation",
         options: ({ queryParams, currentChannelId, currentUserId }: Props) => ({
           refetchQueries: getRefetchQueries(
             queryParams,
             currentChannelId,
             currentUserId
-          )
-        })
+          ),
+        }),
       }
     )
   )(withRouter<FinalProps>(SidebarContainer))
@@ -172,7 +174,7 @@ export default (props: Props) => (
     {({ currentUser }) => (
       <WithProps
         {...props}
-        currentUserId={(currentUser && currentUser._id) || ''}
+        currentUserId={(currentUser && currentUser._id) || ""}
       />
     )}
   </AppConsumer>